requirements.txt
firebase==3.0.1
youtube_dl==2021.6.6
emoji==1.4.2
discord.py==1.7.3
discord==1.7.3
google_api_python_client==2.45.0
python-decouple==3.6
PyNaCl==1.4.0